requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.7.2
backports.zoneinfo==0.2.1
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-celery==3.3.1
djangorestframework==3.14.0
idna==3.4
kombu==3.0.37
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
